pyproject.toml
[project]
name = "stockmarket-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "libmagic>=1.0",
    "notebook>=7.4.7",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.3",
]
